refactor(DataTable): move hooks of manage columns from hooks folder
